greybus: loopback: use a 32-bit count
The count of statistical samples recorded is currently a 64-bit
value.  32 bits is sufficient, and in fact anything more than
that won't work for the do_div() call it's pass to anyway.  So make
the count field be 32 bits.

+++ drivers/staging/greybus/loopback.c
@@ -28,7 +28,7 @@ struct gb_loopback_stats {
 	u32 max;
 	u64 avg;
 	u64 sum;
-	u64 count;
+	u32 count;
 };
 
 struct gb_loopback {
